+++ ads/aqua/extension/ui_handler.py
@@ -8,8 +8,8 @@ from urllib.parse import urlparse
 
 from tornado.web import HTTPError
 
+from ads.aqua.common.decorator import handle_exceptions
 from ads.aqua.data import Tags
-from ads.aqua.decorator import handle_exceptions
 from ads.aqua.extension.base_handler import AquaAPIhandler, Errors
 from ads.aqua.extension.utils import validate_function_parameters
 from ads.aqua.model import ImportModelDetails
