+++ ads/aqua/model.py
@@ -15,21 +15,11 @@ from cachetools import TTLCache
 from huggingface_hub import HfApi, hf_api, snapshot_download
 from oci.data_science.models import JobRun, Model
 
-from ads.aqua import ODSC_MODEL_COMPARTMENT_OCID, logger, utils
+from ads.aqua import ODSC_MODEL_COMPARTMENT_OCID, logger
 from ads.aqua.base import AquaApp, CLIBuilderMixin
-from ads.aqua.constants import (
-    READY_TO_IMPORT_STATUS,
-    TRAINING_METRICS_FINAL,
-    TRINING_METRICS,
-    UNKNOWN_VALUE,
-    VALIDATION_METRICS,
-    VALIDATION_METRICS_FINAL,
-    FineTuningDefinedMetadata,
-)
-from ads.aqua.data import AquaResourceIdentifier, Tags
-from ads.aqua.exception import AquaRuntimeError
-from ads.aqua.training.exceptions import exit_code_dict
-from ads.aqua.utils import (
+from ads.aqua.common import utils
+from ads.aqua.common.errors import AquaRuntimeError
+from ads.aqua.common.utils import (
     LICENSE_TXT,
     MODEL_BY_REFERENCE_OSS_PATH_KEY,
     README,
@@ -42,6 +32,17 @@ from ads.aqua.utils import (
     read_file,
     upload_folder,
 )
+from ads.aqua.constants import (
+    READY_TO_IMPORT_STATUS,
+    TRAINING_METRICS_FINAL,
+    TRINING_METRICS,
+    UNKNOWN_VALUE,
+    VALIDATION_METRICS,
+    VALIDATION_METRICS_FINAL,
+    FineTuningDefinedMetadata,
+)
+from ads.aqua.data import AquaResourceIdentifier, Tags
+from ads.aqua.training.exceptions import exit_code_dict
 from ads.common.auth import default_signer
 from ads.common.extended_enum import ExtendedEnum
 from ads.common.oci_resource import SEARCH_TYPE, OCIResource
@@ -488,9 +489,11 @@ class AquaModelApp(AquaApp):
             if artifact_path != UNKNOWN:
                 model_card = str(
                     read_file(
-                        file_path=f"{artifact_path.rstrip('/')}/config/{README}"
-                        if is_shadow_type
-                        else f"{artifact_path.rstrip('/')}/{README}",
+                        file_path=(
+                            f"{artifact_path.rstrip('/')}/config/{README}"
+                            if is_shadow_type
+                            else f"{artifact_path.rstrip('/')}/{README}"
+                        ),
                         auth=self._auth,
                     )
                 )
