+++ tests/unitary/with_extras/aqua/test_ui_handler.py
@@ -6,16 +6,17 @@
 
 import os
 import unittest
-from unittest.mock import MagicMock, patch
 from importlib import reload
+from unittest.mock import MagicMock, patch
+
+from notebook.base.handlers import IPythonHandler
 from parameterized import parameterized
 
-import ads.config
 import ads.aqua
-from notebook.base.handlers import IPythonHandler
+import ads.config
+from ads.aqua.data import Tags
 from ads.aqua.extension.ui_handler import AquaUIHandler
 from ads.aqua.ui import AquaUIApp
-from ads.aqua.data import Tags
 
 
 class TestDataset:
